// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        BuyerAdmin: {
            readonly id: number;
            name: string;
            licence_no: string;
            user: number;
            readonly user_phone: string;
            /** Format: date-time */
            verified_at?: string | null;
        };
// ...
        SystemUserAdmin: {
            readonly id: number;
            phone: string;
            full_name: string;
            role: components["schemas"]["RoleEnum"];
            scope_level: components["schemas"]["SystemUserAdminScopeLevelEnum"];
            /** Format: int64 */
            scope_id?: number | null;
            is_active?: boolean;
            password?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}